python/packages/corlinman-shadow-tester/pyproject.toml
[project]
name = "corlinman-shadow-tester"
version = "0.1.0"
description = "corlinman ShadowTester (Python port) — shadow-runs medium/high-risk EvolutionProposals against an in-process eval set before they reach the operator queue."
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "corlinman contributors" }]

dependencies = [
    "aiosqlite>=0.20",
    "corlinman-evolution-store",
    "PyYAML>=6.0",
]

[tool.uv.sources]
corlinman-evolution-store = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/corlinman_shadow_tester"]
